feat(projects): add assignment launch preflight
